Moved organisation summary outside registration
