+++ Frontend/CO.CDP.OrganisationApp/Pages/OrganisationOverview.cshtml
@@ -1,6 +1,6 @@
-@page "/registration/organisation-overview"
+@page "/organisation-overview"
 @using CO.CDP.OrganisationApp.Constants
-@model CO.CDP.OrganisationApp.Pages.Registration.OrganisationOverviewModel
+@model OrganisationOverviewModel
 
 @{
     var organisationDetails = Model.OrganisationDetails!;
