+++ Frontend/CO.CDP.OrganisationApp/Pages/OrganisationSelection.cshtml
@@ -40,7 +40,7 @@
                                     </th>
                                     <td class="govuk-table__cell">
                                         <a class="govuk-link govuk-link--no-visited-state"
-                                           href="/registration/organisation-overview?id=@org.Id">Select</a>
+                                           href="/organisation-overview?id=@org.Id">Select</a>
                                     </td>
                                 </tr>
                             }
